split targets for web and slides
Motivated by the different requirements of 'web' and 'slides' processing. In both cases images need to be located at the relative positon stated in markdown files. However, the 'web' creates a folder with the same substructure as the source folder but with alterations to original markdown files so that they can be correctly translated by hugo. This folder shouls also not contain any source or intermediate files as those would be included in the final website. The slides on the other hand are generated from the unaltered markdown files. The solution chosen here was to create image files in the same folder as their source files. This way the files are in the right position for slides generation and can be copied to web folder then needed.