make slides directly dependend on image files
Defining a phony target that bundels image targets in the same folder and setting it as a prerequisite for the pdf file that uses those images did not work. The pdf file would always be rebuild even if nothing had changed. The pdf files now depend directly on the required image files. Also added an empty target for image files that are not generated. This allows them to be used as prerequisites so that dependend targets get rebuild if the imges are modified/replaced manually.